Source file re-ordering after trace for paging performance
