+++ InjectionIII/AppDelegate.swift
@@ -5,7 +5,7 @@
 //  Created by John Holdsworth on 06/11/2017.
 //  Copyright © 2017 John Holdsworth. All rights reserved.
 //
-//  $Id: //depot/ResidentEval/InjectionIII/AppDelegate.swift#89 $
+//  $Id: //depot/ResidentEval/InjectionIII/AppDelegate.swift#90 $
 //
 
 import Cocoa
@@ -376,159 +376,6 @@ class AppDelegate : NSObject, NSApplicationDelegate {
 //            [[NSAlert alertWithError:error] runModal];
 //        }
 //    #endif
-    }
-
-    @IBAction func runXprobe(_ sender: NSMenuItem) {
-        if xprobePlugin == nil {
-            xprobePlugin = XprobePluginMenuController()
-            xprobePlugin.applicationDidFinishLaunching(
-                Notification(name: Notification.Name(rawValue: "")))
-            xprobePlugin.injectionPlugin = unsafeBitCast(self, to: AnyClass.self)
-        }
-        lastConnection?.sendCommand(.xprobe, with: "")
-        windowItem.isHidden = false
-    }
-
-    @IBAction func callOrder(_ sender: NSMenuItem) {
-        lastConnection?.sendCommand(.callOrder, with: nil)
-    }
-
-    @IBAction func fileOrder(_ sender: NSMenuItem) {
-        lastConnection?.sendCommand(.fileOrder, with: nil)
-    }
-
-    func fileOrder(signatures: [String]) {
-        let builder = SwiftEval()
-        builder.projectFile = selectedProject
-
-        guard let projectRoot = selectedProject.flatMap({
-            URL(fileURLWithPath: $0).deletingLastPathComponent().path+"/"
-        }),
-            let (_, logsDir) =
-                try? builder.determineEnvironment(classNameOrFile: "") else {
-            lastConnection?.sendCommand(.log, with:
-                "💉 File ordering not available.")
-            return
-        }
-
-        let tmpfile = NSTemporaryDirectory()+"/eval101"
-        var typesSearched = Set<String>()
-        
-        for signature in signatures {
-            let parts = signature.components(separatedBy: ".")
-            if parts.count < 3 {
-                continue
-            }
-            let typeName = parts[1]
-            if  typesSearched.insert(typeName).inserted,
-                let (_, foundSourceFile) =
-                    try? builder.findCompileCommand(logsDir: logsDir,
-                        classNameOrFile: typeName, tmpfile: tmpfile) {
-                let relativePath = foundSourceFile
-                    .replacingOccurrences(of: projectRoot, with: "")
-                lastConnection?.sendCommand(.log, with: relativePath)
-            }
-        }
-    }
-
-    @objc func injectionGoto(_ pboard: NSPasteboard, userData: NSString,
-                        error: UnsafeMutablePointer<NSString>) {
-        guard pboard.canReadObject(forClasses: [NSString.self], options:nil),
-            let target = pboard.string(forType: .string) else { return }
-
-        let parts = target.components(separatedBy: ".")
-                        .filter { !$0.hasSuffix("init") }
-        let builder = SwiftEval()
-        builder.projectFile = selectedProject
-
-        guard parts.count > 0, let (_, logsDir) =
-            try? builder.determineEnvironment(classNameOrFile: "") else {
-            lastConnection?.sendCommand(.log, with:
-                "💉 Injection Goto service not availble.")
-            return
-        }
-
-        var className: String!, sourceFile: String?
-        let tmpDir = NSTemporaryDirectory()
-
-        for part in parts {
-            let subParts = part.components(separatedBy: " ")
-            className = subParts[0]
-            if let (_, foundSourceFile) =
-                try? builder.findCompileCommand(logsDir: logsDir,
-                        classNameOrFile: className, tmpfile: tmpDir+"/eval101") {
-                sourceFile = foundSourceFile
-                className = subParts.count > 1 ? subParts.last : parts.last
-                break
-            }
-        }
-
-        className = className.replacingOccurrences(of: #"\((\S+).*"#,
-                                                   with: "$1",
-                                                   options: .regularExpression)
-
-        guard sourceFile != nil,
-            let sourceText = try? NSString(contentsOfFile: sourceFile!,
-                                           encoding: String.Encoding.utf8.rawValue),
-            let finder = try? NSRegularExpression(pattern:
-                #"(?:\b(?:var|func|struct|class|enum)\s+|^[+-]\s*(?:\([^)]*\)\s*)?)(\#(className!))\b"#,
-                options: [.anchorsMatchLines]) else {
-            lastConnection?.sendCommand(.log, with:
-                "💉 Unable to find source file for type '\(className!)'.\n")
-            return
-        }
-
-        let match = finder.firstMatch(in: sourceText as String, options: [],
-                                      range: NSMakeRange(0, sourceText.length))
-
-        DispatchQueue.main.async {
-            if let xCode = SBApplication(bundleIdentifier: XcodeBundleID),
-//                xCode.activeWorkspaceDocument.path != nil,
-                let doc = xCode.open(sourceFile!) as? SBObject,
-                doc.selectedCharacterRange != nil,
-                let range = match?.range(at: 1) {
-                doc.selectedCharacterRange =
-                    [NSNumber(value: range.location+1),
-                     NSNumber(value: range.location+range.length)]
-            } else {
-                var numberOfLine = 0, index = 0
-
-                if let range = match?.range(at: 1) {
-                    while index < range.location {
-                        index = NSMaxRange(sourceText
-                                    .lineRange(for: NSMakeRange(index, 0)))
-                        numberOfLine += 1
-                    }
-                }
-
-                guard numberOfLine != 0 else { return }
-
-                var xed = "/usr/bin/xed"
-                if let xcodeURL = self.runningXcodeDevURL {
-                    xed = xcodeURL
-                        .appendingPathComponent("usr/bin/xed").path
-                }
-
-                let script = tmpDir+"/injection_goto.sh"
-                do {
-                    try "\"\(xed)\" --line \(numberOfLine) \"\(sourceFile!)\""
-                        .write(toFile: script, atomically: false, encoding: .utf8)
-                    chmod(script, 0o700)
-
-                    let task = Process()
-                    task.launchPath = "/usr/bin/open"
-                    task.arguments = ["-b", "com.apple.Terminal", script]
-                    task.launch()
-                    task.waitUntilExit()
-                } catch {
-                    NSLog("Failed to write \(script)")
-                }
-            }
-        }
-    }
-
-    @objc func evalCode(_ swift: String) {
-        lastConnection?.sendCommand(.eval, with:swift)
     }
 
     @IBAction func help(_ sender: Any) {
